eTrackingState -> enum class TrackingState in Types.h (issue #15 N5)
Scoped enum, e-prefix dropped. Lives in Types.h because FrameDrawer/Viewer
hold and switch on it while Tracking.h and FrameDrawer.h include each other
— an in-class enum cannot be named from inside that cycle. FrameDrawer's
mState/int locals and DrawTextInfo signature become TrackingState (the
static_cast dance at the copy site disappears); Viewer's trackingStateName
takes the enum; System's public GetTrackingState keeps its int API with one
explicit cast at the copy.

+++ src/System.cc
@@ -337,7 +337,7 @@ mat4f System::Track(Image &im, const double &timestamp)
     mnFramesProcessed.fetch_add(1);
 
     unique_lock<mutex> lock2(mMutexState);
-    mTrackingState = tracker->state_;
+    mTrackingState = static_cast<int>(tracker->state_);
     mTrackedMapPoints = tracker->current_frame_.pts;
     mTrackedKeyPointsUn = tracker->current_frame_.keypoints;
